Moved dashboard context to common folder
