+++ client/src/context/dashboard-context.tsx
@@ -1,19 +1,19 @@
 import { useState, useEffect, createContext, useContext, useCallback } from "react";
-import { GridItem } from "./grid/model/grid-models";
+import { GridItem } from "../core/dashboard/grid/model/grid-models";
 import { v4 as uuidv4 } from "uuid";
-import { GridMetaData } from "./grid/model/grid-models";
-import apiClient from "../../api/ApiClient";
-import { HomeConfig, HomeConfigUtils } from "../../model/HomeConfigState";
-import { useAuth } from "../../context/AuthContext";
-import { ConfigMigration } from "../../lib/version";
-import GridService from "./grid/service/grid-service";
-import { EditingKey, EditModeState } from "./model/EditMode";
-import LoadServerConfig from "./loadServerConfig/load-server-config";
-import { isDefaultView } from "./util/isDefaultView";
-import { useAlert } from "../../feedback/alert/provider/AltertProvider";
-import { AlertVariant } from "../../feedback/alert/model/AlertTypes";
-import { WidgetDefinition, WidgetEnum } from "../../widgets/core/model/widget-type";
-import { WidgetConfigs, Widgets } from "../../widgets/core/model/wigets";
+import { GridMetaData } from "../core/dashboard/grid/model/grid-models";
+import apiClient from "../api/ApiClient";
+import { HomeConfig, HomeConfigUtils } from "../model/HomeConfigState";
+import { useAuth } from "./AuthContext";
+import { ConfigMigration } from "../lib/version";
+import GridService from "../core/dashboard/grid/service/grid-service";
+import { EditingKey, EditModeState } from "../core/dashboard/model/EditMode";
+import LoadServerConfig from "../core/dashboard/loadServerConfig/load-server-config";
+import { isDefaultView } from "../core/dashboard/util/isDefaultView";
+import { useAlert } from "../feedback/alert/provider/AltertProvider";
+import { AlertVariant } from "../feedback/alert/model/AlertTypes";
+import { WidgetDefinition, WidgetEnum } from "../widgets/core/model/widget-type";
+import { WidgetConfigs, Widgets } from "../widgets/core/model/wigets";
 
 type DashboardActions = {
   setWidgets: (widgets: GridItem[]) => void;
