+++ client/src/widgets/core/hooks/useWidgetConfig.ts
@@ -1,4 +1,4 @@
-import { useDashboard } from "../../../core/dashboard/dashboard-context";
+import { useDashboard } from "../../../context/dashboard-context";
 import { WidgetEnum } from "../model/widget-type";
 
 export function useWidgetConfig<TConfig>(widget: WidgetEnum): TConfig | undefined {
