+++ client/src/widgets/laundry-week/configuration/laundryWeekConfiguraiton.tsx
@@ -1,5 +1,5 @@
 import { IoAddCircle } from "react-icons/io5";
-import { useDashboard } from "../../../core/dashboard/dashboard-context";
+import { useDashboard } from "../../../context/dashboard-context";
 import { LaundryWeekConfig } from "../LaundryWeekWidget";
 import { MdDelete } from "react-icons/md";
 import { useState } from "react";
